New method added into the Controller to retrieve new notifications

+++ WebScraper/project/CuokaWebScraperRESTServer/src/main/java/es/sidelab/cuokawebscraperrestserver/beans/User.java
@@ -46,7 +46,7 @@ public class User
     
     @JsonIgnore
     @Column(name = "LAST_NOTIFICATION")
-    private int lastNotification;
+    private long lastNotification;
     
     @JsonIgnore
     @Column(name = "DATE")
@@ -119,7 +119,7 @@ public class User
     public boolean getMan()     { return this.man; }
     public int getPostalCode()  { return this.postalCode; }  
     @JsonIgnore
-    public int getLastNotification()  { return this.lastNotification; }      
+    public long getLastNotification()  { return this.lastNotification; }      
     @JsonIgnore
     public Calendar getRegistrationDate() { return this.registrationDate; }    
     @JsonProperty
@@ -160,7 +160,7 @@ public class User
     @JsonIgnore
     public void setShops(Set<String> shops)   { this.shops = shops; } 
     @JsonIgnore
-    public void setLastNotification(int lastNotification) { this.lastNotification = lastNotification; }
+    public void setLastNotification(long lastNotification) { this.lastNotification = lastNotification; }
     public void setName(String name)          { this.name = name; }   
     public void setEmail(String email)        { this.email = email; }
     public void setPassword(String password)  { this.password = password; }
